+++ RELEASES.md
@@ -14,7 +14,7 @@ Upcoming (7.8)
 - Larva httpRequest parameter is no longer supported
 - Json2XmlValidator input format session key prefix changed from "Json2XmlValidator.inputformat " to "Json2XmlValidator.inputFormat " (capital F)
 - property xml.namespaceAware.default=true by default. When set to false, parsing is done via a DOM source, or namespaces are removed before XSLT transformations.
-
+- Larva uses different correlationId in the format *Test Tool correlation id(${counter})* for each scenario.
 
 7.7
 ---